+++ src/components/atoms/Tab.vue
