+++ src/components/base/BaseTitle.stories.js
@@ -2,7 +2,7 @@ import { storiesOf } from "@storybook/vue";
 
 import BaseTitle from "@components/base/BaseTitle";
 
-storiesOf("Base|BaseComponents|BaseTitle", module).add("BaseTitle", () => ({
+storiesOf("Base|Base UI|BaseTitle", module).add("BaseTitle", () => ({
 	components: { BaseTitle },
 	template: `
 		<BaseTitle>BaseTitle Component</BaseTitle>`,
